==> 05_Score_Weighting_Regression.ipynb <==
import pandas as pd
import certifi
import json
from urllib.request import urlopen
import time

def get_jsonparsed_data(url):
    response = urlopen(url, cafile=certifi.where())
    data = response.read().decode("utf-8")
    return json.loads(data)

url = ("https://financialmodelingprep.com/api/v3/stock-price-change/AAPL?apikey=PHaANSXTwW2zC5hpGFO1uhe8EkPXgio7")
data = get_jsonparsed_data(url)
data[0]

roce_df = pd.read_csv("stock_financials/ROCE_Ratios_recent.csv")
roce_df.drop_duplicates(subset='ticker', inplace=True)
tickers = list(roce_df['ticker'])

main_df = pd.DataFrame()

for ticks in tickers:


    url = (f"https://financialmodelingprep.com/api/v3/stock-price-change/{ticks}?apikey=PHaANSXTwW2zC5hpGFO1uhe8EkPXgio7")
    data = get_jsonparsed_data(url)

    dict = {'ticker' : [],
                '1Y' : [],
                '5Y' : []}
    
    symbol = data[0]['symbol']
    y1 = data[0]['1Y']
    y5 = data[0]['5Y']
  
    
    dict['ticker'].append(symbol)
    dict['1Y'].append(y1)
    dict['5Y'].append(y5)

    df = pd.DataFrame(dict)
    main_df = pd.concat([main_df, df])
    time.sleep(0.05)

main_df.to_csv("Tickers_1_5_Y_retrun.csv")

main_df

stock_scores = pd.read_csv('stock_scores/stock_score_data_recent.csv')
stock_scores

stock_scores = stock_scores.merge(main_df, left_on='ticker', right_on='ticker')

stock_scores.head()

stock_scores.columns

small_df = stock_scores[['ticker', 'coef_rank', 'std_rank', 'roce_rank', '1Y', '5Y']]

small_df.head()

from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score
import numpy as np

def cal_coef(X,y):
    """Function to calculate coefficients or w or theta or b0 & b1"""
    coeffs = np.dot(np.linalg.inv(np.dot(X.T,X)),np.dot(X.T,y)) #
    
    #Similar way to calculate coeffs
    # coeffs = np.linalg.inv(X.T.dot(X)).dot(X.T).dot(y)
    return coeffs
  
def predict_normeq(new_x,coeffs):
    """Function to predict linear regression using normal equation"""
    
    new_y = np.dot(new_x, coeffs)
    return new_y

small_df = small_df.dropna()

X = np.column_stack([np.ones(len(small_df), dtype=np.float32),small_df[['coef_rank', 'roce_rank']]])
y = small_df['5Y']
    

coeffs = cal_coef(X,y) # b0-intercept, b1-slope
slope = coeffs[1]
normeq_preds = predict_normeq(X,coeffs)

    ## Model Evaluation
SSE = sum((y-normeq_preds)**2) # Sum of squared error
SST = sum((y-np.mean(y))**2) # Sum of squared total
R_squared = 1-(SSE/SST) # R Square

std = np.std(y)
    


coeffs[1:] / sum(coeffs[1:])

R_squared

reg_results['ticker'].append(i)
reg_results['coef'].append(slope)
reg_results['r_squared'].append(R_squared)
reg_results['std'].append(std)
