==> 08_email_test.ipynb <==
import smtplib
from email.mime.text import MIMEText
from email.mime.multipart import MIMEMultipart
from datetime import date

def send_email(to_email, subject_body, email_body, from_email, password, smtp_server="smtp.gmail.com", smtp_port=587):
    # Create email content
    msg = MIMEMultipart()
    msg["From"] = from_email
    msg["To"] = to_email
    msg["Subject"] = f"[model results]: {date.today().strftime('%Y-%m-%d')}"
    
    # Add email body
    msg.attach(MIMEText(email_body, "plain"))

    try:
        # Connect to the SMTP server
        server = smtplib.SMTP(smtp_server, smtp_port)
        server.starttls()
        server.login(from_email, password)
        
        # Send email
        server.sendmail(from_email, to_email, msg.as_string())
        print("Email sent successfully!")
    except Exception as e:
        print(f"Failed to send email: {e}")
    finally:
        server.quit()

# Usage example:
to_email = "kristen.anderson08@gmail.com"
subject_body = "Today's model results"
email_body = "Here are the model results for today..."
from_email = "funnyfarmcapital@gmail.com"
password = "wrxf zbqb ycsy bozj"

send_email(to_email, subject_body, email_body, from_email, password)
